+++ RefApp/ViewControllers/AssetListTableViewController.swift
@@ -23,13 +23,10 @@ class AssetListTableViewController: UITableViewController {
         super.viewWillAppear(animated)
         
         self.navigationController?.setNavigationBarHidden(false, animated: false)
-        self.assetsDidLoad([])
     }
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        self.navigationItem.title = NSLocalizedString("Assets", comment: "")
         
         tableView.register(UITableViewCell.self, forCellReuseIdentifier: cellId)
         tableView.tableFooterView = UIView()
@@ -45,6 +42,7 @@ class AssetListTableViewController: UITableViewController {
         button.addTarget(self, action:#selector(handleLogout), for: .touchUpInside)
         button.setTitle(NSLocalizedString("Logout", comment: ""), for: .normal)
         button.setTitleColor(UIColor.white, for: .normal)
+        button.sizeToFit()
         let barButton = UIBarButtonItem(customView: button)
         self.navigationItem.leftBarButtonItem = barButton
     }
@@ -60,9 +58,8 @@ extension AssetListTableViewController {
             return
         }
         
-        let query = "assetType=MOVIE"
+        let query = "assetType=TV_CHANNEL" // MOVIE / TV_CHANNEL
         loadAssets(query: query, environment: environment, endpoint: "/content/asset", method: HTTPMethod.get)
-        
     }
     
     /// Load the assets from the Exposure API
@@ -108,6 +105,45 @@ extension AssetListTableViewController {
         self.tableView.reloadData()
     }
 }
+
+
+// MARK: - TableView Delegate
+extension AssetListTableViewController {
+    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        
+        let asset = assets[indexPath.row]
+        let destinationViewController = PlayerViewController()
+        destinationViewController.environment = StorageProvider.storedEnvironment
+        destinationViewController.sessionToken = StorageProvider.storedSessionToken
+        destinationViewController.channel = asset
+        
+        /// Optional playback properties
+        let properties = PlaybackProperties(autoplay: true,
+                                            playFrom: .bookmark,
+                                            language: .custom(text: "fr", audio: "en"),
+                                            maxBitrate: 300000)
+        
+        destinationViewController.playbackProperties = properties
+        
+        if let type = asset.type {
+            switch type {
+            case "LIVE_EVENT":
+                destinationViewController.playable = ChannelPlayable(assetId: asset.assetId)
+            case "TV_CHANNEL":
+                destinationViewController.playable = ChannelPlayable(assetId: asset.assetId)
+            case "MOVIE":
+                destinationViewController.playable = AssetPlayable(assetId: asset.assetId)
+            default:
+                destinationViewController.playable = AssetPlayable(assetId: asset.assetId)
+                break
+            }
+        }
+        
+        self.navigationController?.pushViewController(destinationViewController, animated: false)
+        tableView.deselectRow(at: indexPath, animated: true)
+    }
+}
+
 
 // MARK: - Actions
 extension AssetListTableViewController {